+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - Renamed `blockbook::websocket::Error`'s `WebsocketError` variant to `Websocket`.
+- Allowed a `blockbook::TransactionSpecific` to be converted into a `bitcoin::Transaction`.
 
 ## 0.1.0